Separate client-side homepage and app view
