+++ client/src/Chat.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import './Chat.css';
+
+class Chat extends Component {
+  render() {
+    return (
+      <div className='chat-container'>
+        <h1>nothing.chat</h1>
+      </div>
+    );
+  }
+}
+
+export default Chat;
