--- server/src/view/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>ForgetChat</title>
-</head>
-<body>
-  <h1>ForgetChat</h1>
-</body>
-</html>
